# dist/api.yml
openapi: 3.0.0
paths: 
...
  /games/AvailableReferees/{gameTime}:
    get:
      tags:
        - "Games"
      summary: "Get all available referees at the time given"
      description: "Get all available referees at the time given"
      parameters:
      - name: "gameTime"
        in: "path"
        description: "The date in which referees need to be avaiable"
        required: true
        schema:
          type: "string"
          example: "2021-11-05 21:05:00"

      responses:
        "404":
          description: "URL not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/ArrayOfReferees"
  /games/AvailableTeams/{gameTime}:
    get:
      tags:
        - "Games"
      summary: "Get all available teams at the time given"
      description: "Get all available teams at the time given"
      operationId: "getAvailableTeams" 
      parameters:
      - name: "gameTime"
        in: "path"
        description: "The date in which referees need to be avaiable"
        required: true
        schema:
          type: "string"
          example: "2021-11-05 21:05:00"

      responses:
        "404":
          description: "URL not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                type: array
                items:
                  type: object
                  allOf:
                    -  $ref: "#/components/schemas/TeamSummary"
  /games/gameid/{game_time}/{home_team_id}/{away_team_id}:
    get:
      tags:
        - "Games"
      summary: "Get game id of given game time and team ids"
      description: "Get game id of given game time and team ids"
      operationId: "getgameid" 
      parameters:
      - name: "game_time"
        in: "path"
        description: "String of game time"
        required: true
        schema:
          type: "string"
          example: "2021-11-05 21:05:00"
      - name: "home_team_id"
        in: "path"
        description: "integer of home team id"
        required: true
        schema:
          type: "integer"
          example: 940
      - name: "away_team_id"
        in: "path"
        description: "integer of away team id"
        required: true
        schema:
          type: "integer"
          example: 939

      responses:
        "405":
          description: "There is no game with the specified input"
        "200":
          description: "successful operation"
          content:
            application/json:  
              schema:
                type: object
                properties:
                  game_id:          
                    type: "integer"
                    minimum: 1
                    example: 6
  /games/addGame:
    post:
      tags:
        - "Games"
      summary: "Add a game to the current stage"
      description: "used for the league representitve to add games to the Superliga league. 
        If the teams ids are not known, do not send them.
       Only teams from the Superliga can be added.
      Only future games can be added. For Future games no results and eventlog should not be given"
      operationId: "addgame"
      requestBody:
        content:
          application/json:  
            schema:
              $ref: "#/components/schemas/FutureGame"

      responses:
        "500":
          description: "Internal Server error"
        "405":
          description: "The teams already have a match in that date & time"
        "406":
          description: "Bad game input. Please check the date or teams"
        "404":
          description: "URL not found"
        "401":
          description: "Privilege Error: The following action is only permitted to league representives Or you have not Logged in first"
        "200":
          description: "The game was updated"
  /games/addScore:
    put:
      tags:
        - "Games"
      summary: "Add a score to a game"
      description: "used for the league representitve to add score to games that were played"
      operationId: "addScore"      
      requestBody:
        content:
          application/json:  
            schema:
              type: object
              properties:
                game_id:
                  type: "integer"
                  example: 1
                score:
                  type: "string"
                  example: "2-1"
      responses:
        "500":
          description: "Internal Server error"
        "404":
          description: "URL not found"
        "401":
          description: "Privilege Error: The following action is only permitted to league representives"
        "200":
          description: "The game was updated"
  /games/addEvent:
    post:
      tags:
        - "Games"
      summary: "Add an event to a game"
      description: "used for the league representitve to add events to games that were played. If the event alreay exist it will be updates with the new event description, else
      it will be created as a new event"
      operationId: "addEvent"      
      requestBody:
        content:
          application/json:  
            schema:
              type: object
              properties:
                game_id:
                  type: "integer"
                  example: 1
                eventlog:
                  $ref: "#/components/schemas/EventLog"

      responses:
        "500":
            description: "Internal Server error"
        "404":
          description: "URL not found"
        "401":
          description: "Privilege Error: The following action is only permitted to league representives"
        "200":
          description: "The game was updated"
...
  /players/id/{playerId}:
    get:
      tags:
        - "Player"
      summary: "used for player card in personal page"
      description: "Get the full information about the player by using the player id"
      operationId: "getPlayerByStageId" 
      parameters:
      - name: "playerId"
        in: "path"
        description: "ID of player to return"
        required: true
        schema:
          type: "integer"
          minimum: 1
          example: 172104
            
      responses:
        "404":
          description: "URL not found"
        "400":
          description: "Player id not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/Player"
  /players/name/{player_name}:
    get:
      tags:
        - "Player"
      summary: "used for player card in personal page"
      description: "Get the full information about the player by using the player fullname"
      operationId: "getPlayerByName" 
      parameters:
      - name: "player_name"
        in: "path"
        description: "Full name of player to return"
        required: true
        schema:
          type: "string"
          example: "James Morrison"
            
      responses:
        "404":
          description: "URL not found"
        "400":
          description: "Player name not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/Player"
  /players/search/{player_name}:
    get:
      tags:
        - "Search"
      summary: "Used for the search page."
      description: "Get part of the player's information by using the player's fullname. All the players with that
      name will return. If there are none, an empty array will return."
      parameters:
      - name: "player_name"
        in: "path"
        description: "Full name of player to return"
        required: true
        schema:
          type: "string"
          example: "Rune Frantsen"
            
      responses:
        "404":
          description: "URL not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:
                type: array
                items:
                  type: object
                  allOf:
                     - $ref: "#/components/schemas/PlayerPreview"
  /players/search/{player_name}/positionId/{player_position_id}:
    get:
      tags:
        - "Search"
      description: "Get part of the player's information by using the player's fullname and position as filter.
      All the players with that name and position id will return. If there are none, an empty array will return."   
      summary: "Used for the search page"
      parameters:
      - name: "player_name"
        in: "path"
        description: "Full name of player to return"
        required: true
        schema:
          type: "string"
          example: "Rune Frantsen"
      - name: "player_position_id"   
        in: "path"
        description: "Player's position id"
        required: true
        schema:
          type: "integer"
          example: 2

      responses:
        "404":
          description: "URL not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:
                type: array
                items:
                  type: object
                  allOf:
                     - $ref: "#/components/schemas/PlayerPreview"
  /players/search/{player_name}/team/{team_name}:
    get:
      tags:
        - "Search"
      summary: "Used for the search page."
      description: "Get part of the player's information by using the player's fullname and team as filter.
      All the players with that name and team name will return. If there are none, an empty array will return."
      parameters:
      - name: "player_name"
        in: "path"
        description: "Full name of player to return"
        required: true
        schema:
          type: "string"
          example: "Esben Lange"
      - name: "team_name"   
        in: "path"
        description: "Player's team name"
        required: true
        schema:
          type: "string"
          example: "Silkeborg"

      responses:
        "404":
          description: "URL not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:
                type: array
                items:
                  type: object
                  allOf:
                     - $ref: "#/components/schemas/PlayerPreview"        
...
  /teams/teamID/{team_name}:
    get:
      tags:
        - "Team"
      description: "Get ID of the team by using the team name"
      summary: "used in the add games page"
      operationId: "getTeamIDByName" 
      parameters:
      - name: "team_name"
        in: "path"
        description: "Name of team id to return"
        required: true
        schema:
          type: "string"
          example: "Silkeborg"
            
      responses:
        "404":
          description: "URL not found - could be because of bad name"
        "400":
          description: "Team name not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:
                type: object
                properties:
                  team_id:
                    type: "integer"
                    example: 939
...
  /games/currentStageGames:
    get:
      tags:
        - "Games"
      description: "Get the latest games for the current stage"
      summary: "for the current stage page"
            
      responses:
        "404":
          description: "URL not found"

        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:
                type: object
                properties:   
                  latest:
                    $ref: "#/components/schemas/ArrayOfGames"
                  upcoming:
                    $ref: "#/components/schemas/ArrayOfFutureGames"
  /games:
     get:
      tags:
        - "Games"
      description: "Get a list of all the games that exists in the system, future and past games."
      summary: "For the league representive page"
            
      responses:
        "404":
          description: "URL not found"

        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/ArrayOfFutureGames"
components:
  schemas:
...
    TeamSummary:
      type: object
      properties:
        team_id:
          type: integer
          example: 211
        team_name:
          type: "string"
          example: "Renders"     
...
    FutureGame:
      required:
        - game_date
        - game_time
        - home_team
        - away_team
        - stadium
      type: object
      properties:
        gameid:
          type: "integer"
          example: 1
        GameDateTime:
          type: "string"
          example: "2021-10-28 17:00:00"
        HomeTeam:
          type: "string"
          example: "Midtjylland"
        HomeTeamID:
          type: "integer"
          example: 939
        AwayTeam:
          type: "string"
          example: "Silkeborg"
        AwayTeamID:
          type: "integer"
          example: 86
        Stadium:
          type: "string"
          example: "MCH Arena"
        Referee:
          type: "string"
          example: "Denis Shalayev"
    Game:
      required:
        - game_date
        - game_time
        - home_team
        - away_team
        - stadium

      type: object
      properties:
        gameid:
          type: "integer"
          example: 1
        GameDateTime:
          type: "string"
          example: "2021-01-13 19:00:00"
        HomeTeam:
          type: "string"
          example: "Midtjylland"
        HomeTeamID:
          type: "integer"
          example: 939
        AwayTeam:
          type: "string"
          example: "Silkeborg"
        AwayTeamID:
          type: "integer"
          example: 211
        Stadium:
          type: "string"
          example: "MCH Arena"
        Referee:
          type: "string"
          example: "Denis Shalayev"
        Result:
          type: "string"
          example: "3-1"
        event_log:
          $ref: "#/components/schemas/EventLog"
    Referee:
      type: object
      properties:
        name:
          type: "string"
          example: "Denis Shalayev"
    GameEvent:
      description: "An Event in a game. Options are: Red Card, Yellow Card, Foul, Offside, Goal, Injury, Subsitution"
      type: object
      properties:
        EventData:
          type: "string"
          example: "2021-01-13"
        EventTime:
          type: "string"
          example: "20:05:00"
        EventGameTime:
          type: "integer"
          format: "int64"
          example: 65  
        EventDesc:
          type: "string"
          example: "Foul Anders Hoff"
    Player:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 37522564  
        team_id:
          type: "integer"
          format: "int64"
          example: 211  
        position:
          type: "string"
          example: "Midfielder" 
        position_id:
          type: "integer"
          example: 2 
        fullname:
          type: "string"
          example: "James Forest"
        common_name:
          type: "string"
          example: "James Forest"
        nationality:
          type: "string"
          example: "Scotland"  
        birthplace:
          type: "string"
          example: "Prestwick" 
        birthdate:
          type: "string"
          example: "07/07/1991"  
        height:
          type: "string"
          example: "175 cm"  
        weight:
          type: "string"
          example: "90 kg"
        team:
          type: "string"
          example: "Celtic"
        image_path:
          type: "string"
          example: "https://cdn.sportmonks.com/images/soccer/players/8/172104.png"
    PlayerPreview:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 172104  
        team_id:
          type: "integer"
          format: "int64"
          example: 53  
        position:
          type: "string"
          example: "Midfielder" 
        position_id:
          type: "integer"
          example: 2 
        fullname:
          type: "string"
          example: "James Forest"
        common_name: 
          type: "string"
          example: "R. Almenning Jarstein"
        nationality: 
          type: "string"
          example: "Norway"
        birthplace:
          type: "string"
          example: "Norway"
        birthdate: 
          type: "string"
          example: "29/09/1984"
        height:
          type: "string"
          example: "192 cm"
        weight:
          type: "string"
          example: "88 kg"
        team:
          type: "string"
          example: "Celtic"
        image_path:
          type: "string"
          example: "https://cdn.sportmonks.com/images/soccer/players/8/172104.png"
...
    EventLog:
      description: "The Eventlog - contains array of all the events in the game. Used only in past games"
      type: array
      items:
        type: object
        allOf:
          - $ref: "#/components/schemas/GameEvent"
...
    ArrayOfReferees:
      type: array
      items:
        type: object
        allOf:
          - $ref: "#/components/schemas/Referee"
...
    ArrayOfGames:
      type: array
      items:
        type: object
        allOf:
          - $ref: "#/components/schemas/Game"
    ArrayOfFutureGames:
      type: array
      items:
        type: object
        allOf:
          - $ref: "#/components/schemas/FutureGame"